requirements.txt
python-dotenv
requests
beautifulsoup4
lxml
faiss-cpu
numpy
streamlit
pytest
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
langchain-huggingface>=0.1.0
sentence-transformers>=2.2.0